+++ samila/genimage.py
@@ -5,7 +5,7 @@ import gc
 import itertools
 import matplotlib
 import matplotlib.pyplot as plt
-from .functions import _GI_initializer, plot_params_filter, generate_params_filter, save_image_filter
+from .functions import _GI_initializer, plot_params_filter, generate_params_filter, save_params_filter
 from .functions import float_range, save_data_file, save_fig_file, save_fig_buf, save_config_file
 from .functions import load_data, load_config, random_equation_gen, nft_storage_upload
 from .params import *
@@ -149,7 +149,7 @@ class GenerativeImage:
         :type depth: float
         :return: result as dict
         """
-        save_image_filter(self, depth)
+        save_params_filter(self, depth)
         response = save_fig_buf(self.fig, self.depth)
         if not response["status"]:
             return {"status": False, "message": response["message"]}
@@ -167,7 +167,7 @@ class GenerativeImage:
         :type depth: float
         :return: result as dict
         """
-        save_image_filter(self, depth)
+        save_params_filter(self, depth)
         return save_fig_file(self.fig, file_adr, self.depth)
 
     def save_data(self, file_adr='data.json'):